chimera-agent-runtime/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
qdrant-client>=1.7.3
langchain-text-splitters
openai>=1.0.0
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-grpc
docling>=1.0.0
minio
nebula3-python>=3.0.0
sentence-transformers